newest

+++ mts.c
@@ -86,7 +86,7 @@ struct loading* addToLoadingQueue(struct loading *loadingHead, struct loading *l
 
 void* waitForTime(void* arg) //waits for the amount of seconds passed through
 { 
-    int seconds = arg;
+    int seconds = (int)arg;
     printf("Waiting for %d seconds\n",seconds);
     int milliSeconds = (1000*seconds); 
     clock_t startTime = clock();
